refactor(meshvpn): consolidate status monitor and remove redundant logs
- Inline status monitor creation, eliminating separate status.go files
- Remove trace-level RPC logs that added noise without debugging value
- Simplify status change callback to only emit events
