+++ internal/meshvpn/tailscale/provider.go
@@ -14,12 +14,19 @@ import (
 	"github.com/jetkvm/kvm/internal/meshvpn"
 )
 
+// statusMonitorLogger adapts the package logger to the StatusMonitorLogger interface.
+type statusMonitorLogger struct{}
+
+func (l *statusMonitorLogger) Warn(msg string, err error, failures int) {
+	logger.Warn().Err(err).Int("consecutiveFailures", failures).Msg(msg)
+}
+
 type Provider struct {
 	mu            sync.RWMutex
 	version       string
 	tunMode       meshvpn.TUNMode
 	process       *ProcessManager
-	statusMonitor *StatusMonitor
+	statusMonitor *meshvpn.StatusMonitor
 	httpClient    meshvpn.HTTPClient
 	versionClient *VersionClient
 	connectCancel context.CancelFunc
@@ -407,7 +414,11 @@ func (p *Provider) StartStatusMonitor(ctx context.Context, onChange meshvpn.Stat
 		p.statusMonitor.Stop()
 	}
 
-	p.statusMonitor = NewStatusMonitor(p, onChange)
+	p.statusMonitor = meshvpn.NewStatusMonitor(meshvpn.StatusMonitorConfig{
+		Provider: p,
+		OnChange: onChange,
+		Logger:   &statusMonitorLogger{},
+	})
 	go p.statusMonitor.Start(ctx)
 }
